Use an alternate setup.py file for wheels
